Fix UE not opening, Added open inventory with BP
Fix UE not being able to open, was some exsample code withing that caused it.

Added the option to open inventory widget with "tab" and fixed Error within logic within the BP code in the widgets. Still Something wrong with refresh inventory within the inventory WBP and pressing current test item will crash game, might be because the item is empty or because of something else within the itemobject code.